fix 386 build
